requirements_pip.txt
acvl-utils==0.2.5
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster==0.7.16
anatcl @ git+https://github.com/carloalbertobarbano/AnatCL.git@afd850510288d41d1b65cf28694a537053c0c0fa
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
Automat==20.2.0
babel==2.17.0
backcall==0.2.0
batchgenerators==0.25.1
batchgeneratorsv2==0.2.3
bcrypt==3.2.0
beautifulsoup4==4.13.3
bids-validator==1.14.7.post0
bidscoin==4.3.0
bidsschematools==1.0.4
black==24.2.0
bleach==6.2.0
blinker==1.4
blosc2==3.3.3
Cerberus==1.3.7
certifi==2020.6.20
cfgv==3.4.0
chardet==4.0.0
citeproc-py==0.8.2
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
commonmark==0.9.0
configobj==5.0.6
connected-components-3d==3.23.0
constantly==15.1.0
contourpy==1.3.1
coverage==7.8.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
dcm2niix==1.0.20220715
decorator==5.2.1
defusedxml==0.7.1
dicom2nifti==2.6.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
docutils==0.20
drmaa==0.7.9
duecredit==0.10.2
dynamic_network_architectures==0.3.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fail2ban==0.11.2
fastjsonschema==2.21.1
fft-conv-pytorch==1.2.0
filelock==3.6.0
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
git-filter-repo==2.47.0
graphviz==0.20.3
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
identify==2.6.9
idna==3.3
imagecodecs==2025.3.30
imageio==2.37.0
imagesize==1.4.1
importlib-metadata==4.6.4
importlib_resources==6.5.2
incremental==21.3.0
iniconfig==2.1.0
iotop==0.6
ipython==8.12.3
isort==5.13.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
linecache2==1.0.0
llvmlite==0.44.0
looseversion==1.3.0
lxml==5.3.1
m2r2==0.3.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
miutil==0.13
mock==5.1.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiecho==0.30
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.9.2
netifaces==0.11.0
networkx==3.4.2
nibabel==5.3.2
nnunetv2==2.6.0
nodeenv==1.9.1
numba==0.61.2
numexpr==2.10.2
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
packaging==24.2
pandas==1.5.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
pep517==0.13.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.2.0
pip-api==0.0.34
pipreqs==0.5.0
platformdirs==2.5.1
plette==2.1.0
pluggy==1.5.0
pre-commit==3.6.0
PrettyPrintout==1.0.4
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.11.1
pydantic_core==2.33.0
pydicom==3.0.1
pydot==2.0.0
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.2.3
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pytest==7.4.0
pytest-console-scripts==1.4.0
pytest-cov==4.1.0
pytest-order==1.2.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gdcm==3.0.24.1
python-magic==0.4.24
pytz==2025.2
PyYAML==5.4.1
pyzmq==26.3.0
recommonmark==0.7.0
referencing==0.36.2
requests==2.25.1
requirementslib==3.0.0
rpds-py==0.24.0
rpl==1.12
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
simpleitk==2.5.0
six==1.16.0
snowballstemmer==2.2.0
sos==4.5.6
soupsieve==2.6
Sphinx==7.2.0
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-apidoc==0.5.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxemoji==0.3.0
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.14.0
systemd-python==234
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torch-geometric==2.6.1
tornado==6.4.2
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
trash-cli==0.17.1.14
triton==3.3.0
tslearn==0.6.3
Twisted==22.1.0
typing-inspection==0.4.0
typing_extensions==4.13.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
xkit==0.0.0
yacs==0.1.8
yarg==0.1.9
yarl==1.20.0
zipp==1.0.0
zope.interface==5.4.0
